requirements.txt
Flask>=2.0
requests
python-dotenv
google-generativeai
google-genai
elevenlabs